refactor: Declare `AwsAccountReadiness` interface
