+++ src/bin/commands/aws-cdk-version.ts
@@ -1,5 +1,5 @@
 import { LibraryInfo } from "../../constants/library-info";
-import type { CliCommandResponse } from "../../types/command";
+import type { CliCommandResponse } from "../../types/cli";
 
 export const awsCdkVersionCommand = (): CliCommandResponse => {
   return Promise.resolve(LibraryInfo.AWS_CDK_VERSIONS);
